+++ cmd/smpe_lint/main.go
@@ -15,7 +15,7 @@ import (
 )
 
 var (
-	version = "v1.0.0"
+	version = "v1.0.1"
 	commit  = "unknown"
 )