+++ cmd/smpe_outl/main.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	version = "v1.0.0"
+	version = "v1.0.1"
 	commit  = "unknown"
 )